pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "instacart-recommender"
version = "1.0.0"
description = "Production-ready Instacart recommendation system with ML models and API"
authors = [{name = "Your Team", email = "team@company.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
keywords = ["machine-learning", "recommendation-system", "instacart", "api", "streamlit"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "streamlit>=1.25.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "tensorflow>=2.13.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
]

[project.urls]
Homepage = "https://github.com/yourorg/instacart-recommender"
Documentation = "https://github.com/yourorg/instacart-recommender/blob/main/README.md"
Repository = "https://github.com/yourorg/instacart-recommender"
Issues = "https://github.com/yourorg/instacart-recommender/issues"

[project.scripts]
instacart-train = "src.main:main"
instacart-serve = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 127
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 127

[tool.pytest.ini_options]
testpaths = ["tests", "."]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"